+++ src/features/login/LoginForm.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import useLogin from "../../hooks/login/useLogin";
 import { Link } from "react-router-dom";
-import styled from "styled-components";
 import naverLogo from "../../assets/imgs/login/네이버로고.png";
 import googleLogo from "../../assets/imgs/login/google-plus.png";
 import kakaoLogo from "../../assets/imgs/login/kakao-talk.png";
@@ -9,6 +8,7 @@ import museoLogo from "../../assets/imgs/museoLogo/임시 로고.png";
 import falseVisibleEyes from "../../assets/imgs/login/invisible_gray.png";
 import trueVisibleEyes from "../../assets/imgs/login/eye_gray.png";
 import Swal from "sweetalert2";
+import * as LoginStyle from "../login/LoginStyle";
 
 function LoginForm() {
   //react-query
@@ -82,16 +82,16 @@ function LoginForm() {
   };
 
   return (
-    <StLogin>
-      <StLinkBox>
+    <LoginStyle.StLogin>
+      <LoginStyle.StLinkBox>
         <Link to="/">
           <img src={museoLogo} alt="museoLogo" />
         </Link>
-      </StLinkBox>
+      </LoginStyle.StLinkBox>
 
-      <StEmailInputBox>
+      <LoginStyle.StEmailInputBox>
         <label>이메일</label>
-        <StEmailInputWrap>
+        <LoginStyle.StEmailInputWrap>
           <input
             type="email"
             name="email"
@@ -99,14 +99,14 @@ function LoginForm() {
             style={{ borderColor: !emailMsg ? "#dddddd" : "red" }}
           />
           <div>{emailMsg}</div>
-        </StEmailInputWrap>
-      </StEmailInputBox>
+        </LoginStyle.StEmailInputWrap>
+      </LoginStyle.StEmailInputBox>
 
-      <StPwInputBox>
+      <LoginStyle.StPwInputBox>
         <label>비밀번호</label>
-        <StPwInputWrap>
+        <LoginStyle.StPwInputWrap>
           {!pwVisible ? (
-            <StPwInputImgWrap>
+            <LoginStyle.StPwInputImgWrap>
               <input
                 type="password"
                 name="password"
@@ -122,9 +122,9 @@ function LoginForm() {
               <div onClick={visibleChangeHandler}>
                 <img src={falseVisibleEyes} alt="invisibleEyes" />
               </div>
-            </StPwInputImgWrap>
+            </LoginStyle.StPwInputImgWrap>
           ) : (
-            <StPwInputImgWrap>
+            <LoginStyle.StPwInputImgWrap>
               <input
                 type="text"
                 name="password"
@@ -136,225 +136,42 @@ function LoginForm() {
               <div onClick={visibleChangeHandler}>
                 <img src={trueVisibleEyes} alt="trueVisibleEyes" />
               </div>
-            </StPwInputImgWrap>
+            </LoginStyle.StPwInputImgWrap>
           )}
-        </StPwInputWrap>
-        <StPwWarning>{pwMsg}</StPwWarning>
-      </StPwInputBox>
+        </LoginStyle.StPwInputWrap>
+        <LoginStyle.StPwWarning>{pwMsg}</LoginStyle.StPwWarning>
+      </LoginStyle.StPwInputBox>
 
-      <StLoginBtn onClick={loginHandler}>로그인</StLoginBtn>
+      <LoginStyle.StLoginBtn onClick={loginHandler}>
+        로그인
+      </LoginStyle.StLoginBtn>
 
-      <StSnsBox>
+      <LoginStyle.StSnsBox>
         <div>SNS로 간편하게 시작하기</div>
 
-        <StSnsBtnWrap>
-          <GoogleLogoDiv onClick={socialLoginBtn}>
+        <LoginStyle.StSnsBtnWrap>
+          <LoginStyle.GoogleLogoDiv onClick={socialLoginBtn}>
             <img
               src={googleLogo}
               alt="googleLogo"
               style={{ width: "30px", height: "30px" }}
             />
-          </GoogleLogoDiv>
+          </LoginStyle.GoogleLogoDiv>
           <div onClick={socialLoginBtn}>
             <img src={kakaoLogo} alt="kakaoLogo" />
           </div>
           <div onClick={socialLoginBtn}>
             <img src={naverLogo} alt="naverLogo" />
           </div>
-        </StSnsBtnWrap>
-      </StSnsBox>
+        </LoginStyle.StSnsBtnWrap>
+      </LoginStyle.StSnsBox>
 
-      <StRegisterLink>
+      <LoginStyle.StRegisterLink>
         <div>아직 회원이 아니시라면</div>
-        <StLink to="/register">회원가입</StLink>
-      </StRegisterLink>
-    </StLogin>
+        <LoginStyle.StLink to="/register">회원가입</LoginStyle.StLink>
+      </LoginStyle.StRegisterLink>
+    </LoginStyle.StLogin>
   );
 }
 
 export default LoginForm;
-
-const StLogin = styled.form`
-  font-family: "SpoqaHanSansNeo-Regular";
-  width: 616px;
-  height: 779px;
-  background-color: white;
-  border-radius: 10px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  box-shadow: 0px 5px 20px 0px rgba(148, 148, 148, 0.25);
-`;
-
-const StLinkBox = styled.div`
-  background-color: white;
-  width: 217px;
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 72px;
-`;
-
-const StEmailInputBox = styled.div`
-  width: 416px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  margin: 69px 0px 26px;
-
-  label {
-    color: #242424;
-    font-size: 15px;
-    font-weight: bold;
-  }
-`;
-
-const StEmailInputWrap = styled.div`
-  width: 416px;
-  height: 48px;
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-bottom: 5px;
-  }
-
-  div {
-    color: #f65959;
-  }
-`;
-
-const StPwInputBox = styled.div`
-  width: 416px;
-  display: flex;
-  flex-direction: column;
-
-  label {
-    color: #242424;
-    font-size: 15px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-`;
-
-const StPwInputWrap = styled.div`
-  width: 416px;
-  height: 48px;
-
-  div {
-    color: #f65959;
-  }
-`;
-
-const StPwInputImgWrap = styled.div`
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    position: absolute;
-  }
-
-  img {
-    width: 20px;
-    height: 20px;
-    transform: translate(385px, 12px);
-    cursor: pointer;
-  }
-`;
-
-const StPwWarning = styled.div`
-  color: #f65959;
-  height: 9px;
-`;
-
-const StLoginBtn = styled.button`
-  font-family: "SpoqaHanSansNeo-Regular";
-  background-color: white;
-  color: #171717;
-  width: 416px;
-  height: 65px;
-  border: 1px solid gray;
-  border-radius: 50px;
-  font-size: 15px;
-  font-weight: bold;
-  margin: 44px 0px 72px;
-  cursor: pointer;
-
-  &:hover {
-    border: none;
-    background-color: #242424;
-    color: white;
-  }
-`;
-
-const StSnsBox = styled.div`
-  width: 195px;
-  height: 86px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 23px;
-
-  div {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 15px;
-    font-weight: bold;
-  }
-`;
-
-const StSnsBtnWrap = styled.div`
-  display: flex;
-  gap: 16px;
-
-  div {
-    width: 48px;
-    height: 48px;
-    border-radius: 50%;
-    cursor: pointer;
-  }
-
-  img {
-    width: 48px;
-    height: 48px;
-    border-radius: 50%;
-    cursor: pointer;
-  }
-`;
-
-const GoogleLogoDiv = styled.div`
-  background-color: #f65959;
-  /* border: 1px solid rgba(148, 148, 148, 0.25); */
-`;
-
-const StRegisterLink = styled.div`
-  font-family: "SpoqaHanSansNeo-Regular";
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 50px;
-  gap: 12px;
-
-  div {
-    font-size: 16px;
-    color: #5a5a5a;
-  }
-`;
-
-const StLink = styled(Link)`
-  font-size: 16px;
-  color: #3360ff;
-`;
