+++ src/features/mypage/UserProfile.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from "react";
 import { useGetUserProfile } from "../../hooks/mypage/useGetUserProfile";
 import UpdateUserProfileModal from "./UpdateUserProfileModal";
 import UpdateModalBlackBg from "./UpdateModalBlackBg";
-import styled from "styled-components";
 import AlarmContainer from "./AlarmContainer";
+import * as UserProfileStyle from "../mypage/UserProfileStyle";
 import palette from "../../assets/imgs/mypage/palette_gradient.png";
 import setting from "../../assets/imgs/mypage/gear_gray.png";
 import { usePatchRole } from "../../hooks/mypage/usePatchRole";
@@ -76,15 +76,21 @@ function UserProfile() {
 
   return (
     <>
-      <StUserProfileBox>
-        <ProfileImg src={userProfile?.profileImg} alt="userProfileImg" />
+      <UserProfileStyle.StUserProfileBox>
+        <UserProfileStyle.ProfileImg
+          src={userProfile?.profileImg}
+          alt="userProfileImg"
+        />
 
-        <StEditBtnWrap>
-          <UpdateBtn onClick={openSettingHandler}>
+        <UserProfileStyle.StEditBtnWrap>
+          <UserProfileStyle.UpdateBtn onClick={openSettingHandler}>
             <img src={setting} alt="setting" />
-          </UpdateBtn>
+          </UserProfileStyle.UpdateBtn>
           {openSet && (
-            <StSettingBtn ref={ref} onMouseDown={e => e.stopPropagation()}>
+            <UserProfileStyle.StSettingBtn
+              ref={ref}
+              onMouseDown={e => e.stopPropagation()}
+            >
               <div onClick={updateUserProfileHandler}>프로필 수정</div>
               <button
                 onClick={changeRoleHandler}
@@ -92,27 +98,31 @@ function UserProfile() {
               >
                 작가 신청
               </button>
-            </StSettingBtn>
+            </UserProfileStyle.StSettingBtn>
           )}
-        </StEditBtnWrap>
+        </UserProfileStyle.StEditBtnWrap>
 
-        <StUserNameWrap>
-          <StInfoUserName>{userProfile?.nickname}</StInfoUserName>
+        <UserProfileStyle.StUserNameWrap>
+          <UserProfileStyle.StInfoUserName>
+            {userProfile?.nickname}
+          </UserProfileStyle.StInfoUserName>
           {userProfile?.role === "UR02" ? (
-            <StArtistMark>
+            <UserProfileStyle.StArtistMark>
               <img src={palette} alt="palette" />
-            </StArtistMark>
+            </UserProfileStyle.StArtistMark>
           ) : null}
-        </StUserNameWrap>
+        </UserProfileStyle.StUserNameWrap>
 
-        <StUserInfoIntro>
-          <InfoIntro>{userProfile?.introduction}</InfoIntro>
-        </StUserInfoIntro>
+        <UserProfileStyle.StUserInfoIntro>
+          <UserProfileStyle.InfoIntro>
+            {userProfile?.introduction}
+          </UserProfileStyle.InfoIntro>
+        </UserProfileStyle.StUserInfoIntro>
 
-        <Line></Line>
+        <UserProfileStyle.Line></UserProfileStyle.Line>
 
         <AlarmContainer />
-      </StUserProfileBox>
+      </UserProfileStyle.StUserProfileBox>
 
       {/* 유저 프로필 수정을 위한 모달 open */}
       {openModal && <UpdateUserProfileModal setOpenModal={setOpenModal} />}
@@ -123,158 +133,3 @@ function UserProfile() {
 }
 
 export default UserProfile;
-
-const StUserProfileBox = styled.div`
-  background-color: #ffffff;
-  border-radius: 10px;
-  width: 430px;
-  height: 822px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 74px;
-  margin-left: 20px;
-`;
-
-const StEditBtnWrap = styled.div`
-  width: 450px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  padding: 10px 20px 0px;
-  position: relative;
-`;
-
-const UpdateBtn = styled.button`
-  background-color: white;
-  width: 41px;
-  height: 41px;
-  border-radius: 50%;
-  cursor: pointer;
-
-  img {
-    width: 25px;
-    height: 25px;
-  }
-`;
-
-const StSettingBtn = styled.div`
-  margin-top: 40px;
-  position: absolute;
-  width: 135px;
-  height: 64px;
-  background-color: #ffffff;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 3px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  div {
-    font-family: "SpoqaHanSansNeo-Regular";
-    font-size: 12px;
-    color: #3c3c3c;
-    width: 135px;
-    height: 32px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-    border-radius: 3px;
-    border-bottom: 1px solid #dddddd;
-
-    &:hover {
-      background-color: #3c3c3c;
-      color: #f65959;
-    }
-  }
-
-  button {
-    border: none;
-    background-color: inherit;
-    font-family: "SpoqaHanSansNeo-Regular";
-    font-size: 12px;
-    color: #3c3c3c;
-    width: 135px;
-    height: 32px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-    border-radius: 3px;
-
-    button:disabled {
-      cursor: default;
-      background-color: "#EEEEEE";
-      color: "#7E7E7E";
-    }
-
-    &:hover:not(:disabled) {
-      background-color: #3c3c3c;
-      color: #f65959;
-    }
-  }
-`;
-
-const StUserNameWrap = styled.div`
-  width: 400px;
-  height: 28px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 42px;
-  gap: 8px;
-`;
-
-const StInfoUserName = styled.div`
-  font-family: "Montserrat", sans-serif;
-  font-size: 32px;
-  font-weight: 600;
-`;
-
-const StArtistMark = styled.div`
-  /* background-color: #808080bd; */
-  width: 28px;
-  height: 28px;
-  border-radius: 50%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  img {
-    width: 25px;
-    height: 25px;
-  }
-`;
-
-const StUserInfoIntro = styled.div`
-  font-size: 16px;
-  height: 90px;
-  width: 250px;
-  margin-top: 35px;
-`;
-
-const InfoIntro = styled.div`
-  font-family: "SpoqaHanSansNeo-Regular";
-  font-size: 16px;
-  color: #3c3c3c;
-  /* font-weight: 400; */
-  line-height: 25px;
-  word-break: break-all;
-  text-align: center;
-`;
-
-const ProfileImg = styled.img`
-  width: 150px;
-  height: 150px;
-  border-radius: 50%;
-  transform: translate(0%, -50%);
-  position: absolute;
-`;
-
-const Line = styled.div`
-  border-top: 1px solid #cccccc;
-  width: 392px;
-  height: 32px;
-`;
