+++ src/features/register/RegisterForm.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { useRegister } from "../../hooks/register/useRegister";
 import { useEmailConfirm } from "../../hooks/register/useEmailConfirm";
-import styled from "styled-components";
 import { useEmailAuthSend } from "../../hooks/register/useEmailAuthSend";
 import { useEmailAuthConfirm } from "../../hooks/register/useEmailAuthComfirm";
+import * as RegisterFormStyle from "../register/RegisterFormStyle";
 import museoLogo from "../../assets/imgs/museoLogo/임시 로고.png";
 import falseVisibleEyes from "../../assets/imgs/login/invisible_gray.png";
 import trueVisibleEyes from "../../assets/imgs/login/eye_gray.png";
@@ -216,19 +216,19 @@ function RegisterForm() {
   };
 
   return (
-    <StRegister>
-      <StRegisterWrap>
-        <StLinkBox>
+    <RegisterFormStyle.StRegister>
+      <RegisterFormStyle.StRegisterWrap>
+        <RegisterFormStyle.StLinkBox>
           <Link to="/">
             <img src={museoLogo} alt="museoLogo" />
           </Link>
-        </StLinkBox>
+        </RegisterFormStyle.StLinkBox>
 
-        <StEmailWrap>
-          <StEmailInputBox>
+        <RegisterFormStyle.StEmailWrap>
+          <RegisterFormStyle.StEmailInputBox>
             <label>이메일</label>
 
-            <StEmailInputBtn>
+            <RegisterFormStyle.StEmailInputBtn>
               <input
                 type="email"
                 name="email"
@@ -249,17 +249,17 @@ function RegisterForm() {
               >
                 중복 확인
               </button>
-            </StEmailInputBtn>
+            </RegisterFormStyle.StEmailInputBtn>
 
-            <StEmailInputWarning>
+            <RegisterFormStyle.StEmailInputWarning>
               <div>{emailMsg || warningMsg || emailConfirmMsg}</div>
-            </StEmailInputWarning>
-          </StEmailInputBox>
+            </RegisterFormStyle.StEmailInputWarning>
+          </RegisterFormStyle.StEmailInputBox>
 
-          <StEmailValidationBox>
+          <RegisterFormStyle.StEmailValidationBox>
             <label>이메일 인증</label>
 
-            <StEmailAuthBox>
+            <RegisterFormStyle.StEmailAuthBox>
               <input
                 type="text"
                 value={code}
@@ -268,9 +268,9 @@ function RegisterForm() {
               />
 
               {showTimer && (
-                <StCount>
+                <RegisterFormStyle.StCount>
                   {minutes}:{seconds < 10 ? `0${seconds}` : seconds}
-                </StCount>
+                </RegisterFormStyle.StCount>
               )}
               <button
                 onClick={emailAuthSendHandler}
@@ -283,21 +283,21 @@ function RegisterForm() {
               >
                 인증번호 발송
               </button>
-            </StEmailAuthBox>
-            <StEmailAuthBtn onClick={emailAuthConfirmHandler}>
+            </RegisterFormStyle.StEmailAuthBox>
+            <RegisterFormStyle.StEmailAuthBtn onClick={emailAuthConfirmHandler}>
               확인
-            </StEmailAuthBtn>
-            <StEmailAuthWarning>
+            </RegisterFormStyle.StEmailAuthBtn>
+            <RegisterFormStyle.StEmailAuthWarning>
               <div>{emailAuthMsg}</div>
-            </StEmailAuthWarning>
-          </StEmailValidationBox>
-        </StEmailWrap>
+            </RegisterFormStyle.StEmailAuthWarning>
+          </RegisterFormStyle.StEmailValidationBox>
+        </RegisterFormStyle.StEmailWrap>
 
-        <StPwWrap>
-          <StPwContainer>
+        <RegisterFormStyle.StPwWrap>
+          <RegisterFormStyle.StPwContainer>
             <label>비밀번호</label>
             {!pwVisible ? (
-              <StPwInputImgWrap>
+              <RegisterFormStyle.StPwInputImgWrap>
                 <input
                   type="password"
                   name="password"
@@ -314,9 +314,9 @@ function RegisterForm() {
                 <div onClick={visibleChangeHandler}>
                   <img src={falseVisibleEyes} alt="invisibleEyes" />
                 </div>
-              </StPwInputImgWrap>
+              </RegisterFormStyle.StPwInputImgWrap>
             ) : (
-              <StPwInputImgWrap>
+              <RegisterFormStyle.StPwInputImgWrap>
                 <input
                   type="text"
                   name="password"
@@ -327,12 +327,14 @@ function RegisterForm() {
                 <div onClick={visibleChangeHandler}>
                   <img src={trueVisibleEyes} alt="trueVisibleEyes" />
                 </div>
-              </StPwInputImgWrap>
+              </RegisterFormStyle.StPwInputImgWrap>
             )}
-            <StPwInputWarning>{pwMsg}</StPwInputWarning>
-          </StPwContainer>
+            <RegisterFormStyle.StPwInputWarning>
+              {pwMsg}
+            </RegisterFormStyle.StPwInputWarning>
+          </RegisterFormStyle.StPwContainer>
 
-          <StPwConfirmContainer>
+          <RegisterFormStyle.StPwConfirmContainer>
             <label>비밀번호 확인</label>
             <div>
               <input
@@ -350,11 +352,13 @@ function RegisterForm() {
               />
             </div>
 
-            <StPwCheckWarning>{checkPwMsg}</StPwCheckWarning>
-          </StPwConfirmContainer>
-        </StPwWrap>
+            <RegisterFormStyle.StPwCheckWarning>
+              {checkPwMsg}
+            </RegisterFormStyle.StPwCheckWarning>
+          </RegisterFormStyle.StPwConfirmContainer>
+        </RegisterFormStyle.StPwWrap>
 
-        <StNickNameBox>
+        <RegisterFormStyle.StNickNameBox>
           <label>닉네임</label>
 
           <div>
@@ -365,369 +369,18 @@ function RegisterForm() {
               placeholder="2글자 이상 8글자 이하 입력"
               style={{ borderColor: nicknameMsg ? "red" : "#dddddd" }}
             />
-            <StNickNameWarning>{nicknameMsg}</StNickNameWarning>
+            <RegisterFormStyle.StNickNameWarning>
+              {nicknameMsg}
+            </RegisterFormStyle.StNickNameWarning>
           </div>
-        </StNickNameBox>
+        </RegisterFormStyle.StNickNameBox>
 
-        <StRegisterBtn onClick={registerHandler}>가입하기</StRegisterBtn>
-      </StRegisterWrap>
-    </StRegister>
+        <RegisterFormStyle.StRegisterBtn onClick={registerHandler}>
+          가입하기
+        </RegisterFormStyle.StRegisterBtn>
+      </RegisterFormStyle.StRegisterWrap>
+    </RegisterFormStyle.StRegister>
   );
 }
 
 export default RegisterForm;
-
-const StRegister = styled.div`
-  font-family: "SpoqaHanSansNeo-Regular";
-  background-color: white;
-  border-radius: 10px;
-  width: 616px;
-  height: 852px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  box-shadow: 0px 5px 20px 0px rgba(148, 148, 148, 0.25);
-`;
-
-const StRegisterWrap = styled.div`
-  width: 416px;
-  height: 700;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const StLinkBox = styled.div`
-  background-color: white;
-  width: 217px;
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 52px 0px 36px 0px;
-`;
-
-const StEmailWrap = styled.div`
-  width: 416px;
-  height: 260px;
-  display: flex;
-  flex-direction: column;
-`;
-
-const StEmailInputBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-
-  label {
-    height: 12px;
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 329px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-right: 8px;
-  }
-
-  button {
-    /* background-color: white; */
-    /* color: #3c3c3c; */
-    border: 1px solid;
-    border-radius: 5px;
-    width: 79px;
-    height: 44px;
-    font-size: 12px;
-    font-family: "SpoqaHanSansNeo-Regular";
-    cursor: pointer;
-  }
-`;
-
-const StEmailInputBtn = styled.div`
-  width: 416px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 8px;
-`;
-
-const StEmailInputWarning = styled.div`
-  width: 416px;
-  height: 9px;
-  display: flex;
-  align-items: center;
-  color: #f65959;
-  margin-bottom: 15px;
-
-  div {
-    font-size: 12px;
-  }
-`;
-
-const StEmailValidationBox = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  label {
-    height: 12px;
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-`;
-
-const StEmailAuthBox = styled.div`
-  width: 436px;
-  display: flex;
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 299px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-right: 8px;
-    position: relative;
-
-    &::placeholder {
-      color: #cccccc;
-      font-family: "Montserrat", sans-serif;
-      font-size: 14px;
-    }
-  }
-
-  button {
-    background-color: white;
-    color: #3c3c3c;
-    border: 1px solid #3c3c3c;
-    border-radius: 5px;
-    width: 109px;
-    height: 44px;
-    font-size: 12px;
-    font-family: "SpoqaHanSansNeo-Regular";
-    cursor: pointer;
-    margin-bottom: 8px;
-  }
-`;
-
-const StCount = styled.div`
-  font-family: "Montserrat", sans-serif;
-  font-size: 12px;
-  color: #f65959;
-  width: 20px;
-  height: 20px;
-  position: absolute;
-  transform: translate(260px, 15px);
-`;
-
-const StEmailAuthBtn = styled.button`
-  background-color: white;
-  color: #3c3c3c;
-  border: 1px solid #3c3c3c;
-  border-radius: 5px;
-  width: 416px;
-  height: 44px;
-  font-size: 12px;
-  font-family: "SpoqaHanSansNeo-Regular";
-  cursor: pointer;
-  margin-bottom: 8px;
-`;
-
-const StEmailAuthWarning = styled.div`
-  width: 416px;
-  height: 9px;
-  display: flex;
-  align-items: center;
-  color: #f65959;
-  margin-bottom: 15px;
-
-  div {
-    font-size: 12px;
-  }
-`;
-
-const StPwWrap = styled.div`
-  width: 416px;
-  height: 208px;
-  display: flex;
-  flex-direction: column;
-`;
-
-const StPwContainer = styled.div`
-  width: 416px;
-  height: 104px;
-  display: flex;
-  flex-direction: column;
-
-  label {
-    height: 12px;
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-right: 8px;
-
-    &::placeholder {
-      color: #cccccc;
-      font-family: "Montserrat", sans-serif;
-      font-size: 14px;
-      letter-spacing: 0;
-    }
-  }
-`;
-
-const StPwConfirmContainer = styled.div`
-  width: 416px;
-  height: 104px;
-  display: flex;
-  flex-direction: column;
-
-  label {
-    height: 12px;
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-right: 8px;
-  }
-`;
-
-const StPwInputImgWrap = styled.div`
-  margin-bottom: 26px;
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    position: absolute;
-  }
-
-  img {
-    width: 20px;
-    height: 20px;
-    transform: translate(385px, 12px);
-    cursor: pointer;
-  }
-`;
-
-const StPwInputWarning = styled.div`
-  width: 416px;
-  height: 9px;
-  display: flex;
-  align-items: center;
-  color: #f65959;
-  margin: 8px 0px 15px 0px;
-  font-size: 12px;
-`;
-
-const StPwCheckWarning = styled.div`
-  width: 416px;
-  height: 9px;
-  display: flex;
-  align-items: center;
-  color: #f65959;
-  margin-top: 8px;
-  font-size: 12px;
-`;
-
-const StNickNameBox = styled.div`
-  width: 416px;
-  height: 89px;
-  display: flex;
-  flex-direction: column;
-
-  label {
-    height: 12px;
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-
-  input {
-    font-family: "Montserrat", sans-serif;
-    width: 416px;
-    height: 44px;
-    padding: 10px;
-    border: 1px solid #dddddd;
-    border-radius: 5px;
-    outline: none;
-    font-size: 16px;
-    margin-right: 8px;
-
-    &::placeholder {
-      color: #cccccc;
-      font-family: "Montserrat", sans-serif;
-      font-size: 14px;
-    }
-  }
-`;
-
-const StNickNameWarning = styled.div`
-  width: 416px;
-  height: 9px;
-  display: flex;
-  align-items: center;
-  color: #f65959;
-  margin-top: 8px;
-  font-size: 12px;
-`;
-
-const StRegisterBtn = styled.button`
-  margin-top: 40px;
-  background-color: white;
-  text-decoration: none;
-  font-weight: bold;
-  color: #171717;
-  width: 416px;
-  height: 65px;
-  border-radius: 50px;
-  border: 1px solid #171717;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-family: "SpoqaHanSansNeo-Regular";
-  font-size: 16px;
-  cursor: pointer;
-
-  &:hover {
-    border: none;
-    background-color: #242424;
-    color: white;
-  }
-`;
